Movement Ready and 2 Player test Ready

+++ src/model/Coord.java
@@ -8,7 +8,7 @@ public class Coord {
 
 	    public Coord(){}
 
-	    public Coord(int x,int y) {
+	    public Coord(float x,float y) {
 	        super();
 	        this.x=x;
 	        this.y=y;
